refactor: separate vscode lib from extension and core linter
