+++ package.json
@@ -24,7 +24,7 @@
   "activationEvents": [
     "onLanguage:yaml"
   ],
-  "main": "./out/extension.js",
+  "main": "./out/src/extension/extension.js",
   "contributes": {
     "languages": [
       {
@@ -72,10 +72,10 @@
     }
   },
   "scripts": {
+    "test": "jest --detectOpenHandles --passWithNoTests",
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
-    "pretest": "npm run compile && npm run lint",
     "lint": "eslint src --ext .ts,.js --max-warnings 0",
     "build": "tsc",
     "package": "vsce package",
@@ -88,21 +88,22 @@
     "@semantic-release/github": "github:semantic-release/github",
     "@semantic-release/npm": "github:semantic-release/npm",
     "@semantic-release/release-notes-generator": "github:semantic-release/release-notes-generator",
+    "@types/jest": "^29.5.14",
     "@types/js-yaml": "^4.0.9",
-    "@types/mocha": "^10.0.10",
     "@types/node": "20.x",
     "@types/vscode": "^1.99.0",
     "@typescript-eslint/eslint-plugin": "^8.28.0",
     "@typescript-eslint/parser": "^8.28.0",
-    "@vscode/test-cli": "^0.0.10",
     "@vscode/test-electron": "^2.4.1",
     "@vscode/vsce": "^3.3.2",
     "esbuild": "^0.25.1",
     "esbuild-node-externals": "^1.13.0",
     "eslint": "^9.23.0",
+    "jest": "^29.7.0",
     "npm-run-all": "^4.1.5",
     "semantic-release": "^24.2.3",
-    "typescript": "^5.8.2"
+    "ts-jest": "^29.3.4",
+    "typescript": "^5.8.3"
   },
   "dependencies": {
     "js-yaml": "^4.1.0"
