+++ test/fixtures/valid/map-inputs.yaml
@@ -1,0 +1,12 @@
+version: '2.0'
+workflows:
+  map-inputs:
+    type: direct
+    input:
+      vm_name: null
+      retries: 3
+    tasks:
+      t1:
+        action: std.echo
+        input:
+          output: "VM: <% $.vm_name %> Retries: <% $.retries %>"
